// frontend/src/types/Bowlers.ts
export type Bowlers = {
  bowlerId: number;
  bowlerFirstName: string;
  bowlerMiddleInit: string; //maybe character
  bowlerLastName: string;
  bowlerAddress: string;
  bowlerCity: string;
  bowlerState: string;
  bowlerZip: string;
  bowlerPhoneNumber: string;
  teamId: number;
  teamName: {
    teamName: string;
  };
};